requirements-dev.txt
pyqt5
appdirs
watchdog
python-dotenv
lxml ; sys_platform == 'win32'
pypiwin32 ; sys_platform == 'win32'
py2app ; sys_platform == 'darwin'
pytest
pressenter2exit
balsa
tobool
requests
ismain
pyship
black
wheel
typeguard
sentry-sdk
flake8
mypy
PyQt5-stubs
appdirs-stubs
pref
